feliciano/requirements.txt
Flask
gunicorn
python-dotenv
google.generativeai
Flask-CORS
Werkzeug